requirements.txt
fastapi
httpx
uvicorn
monero
argon2-cffi
PyJWT
